apps/agent/pyproject.toml
[tool.poetry]
name = "maru-agent"
version = "0.1.0"
description = "Shadow coding agent using Claude Agent SDK"
authors = ["Moru <team@moru.io>"]
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
moru = "^0.1.0"
python-dotenv = "^1.0.0"
claude-agent-sdk = "^0.1.0"

[tool.poetry.group.dev.dependencies]
datamodel-code-generator = "^0.28.0"

[tool.poetry.scripts]
generate-protocol = "scripts.generate_protocol:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
